adding room details in database

+++ roomManagementSystem/models.py
@@ -5,8 +5,10 @@ from django.db import models
 class rooms(models.Model):
     hotel_name = models.CharField(max_length=30)
     room_type = models.CharField(max_length=30)
-    room_price = models.IntegerField(max_length=30)
+    room_price = models.FloatField(max_length=30)
     room_image = models.ImageField(upload_to='images')
     room_desc = models.CharField(max_length=30)
+
+
